backend/pyproject.toml
[project]
name = "app-stack"
version = "0.1.0"
description = "EDGE demo project to showcase three applications"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.22",
    "fastapi>=0.110.0",
    "jsonschema>=4.26.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-oci>=0.2.1",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=0.2.0",
    "mcp>=1.25.0",
    "oracledb>=3.4.2",
    "pydantic>=2.6.0",
    "pypdf>=6.7.4",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
